Make a new bc-bot package from the API
