+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules
 config.json
-out
+dist/
 
 config*.json
 !config.sample.json
