+++ Dockerfile
@@ -1,6 +1,6 @@
 FROM node:20
 
-COPY ./out/bundle.js* /bot/
+COPY ./dist/bundle.js* /bot/
 
 WORKDIR /bot/cfg
 CMD ["node", "--enable-source-maps", "/bot/bundle.js"]
