+++ bin/hub/logic/administrationLogic.ts
@@ -2,16 +2,15 @@
  * Copied from the original bot hub, with permission.
  */
 
-import { logger  } from "../../../src/api";
+import { logger, API_Connector,
+API_Character,
+API_Chatroom,
+AnyBotEvent, AnyCharacterEvent, AnyLogicEvent, LogicEvent_Beep, LogicEvent_CharacterEntered, LogicEvent_CharacterLeft, LogicEvent_Message, LogicEvent_RoomForceLeave, LogicEvent_RoomUpdate,
+LogicBase } from "bc-bot";
 import promClient from "prom-client";
 
 import _ from "lodash";
 import fs from "fs";
-import { API_Connector } from "../../../src/apiConnector";
-import { API_Character } from "../../../src/apiCharacter";
-import { API_Chatroom } from "../../../src/apiChatroom";
-import { AnyBotEvent, AnyCharacterEvent, AnyLogicEvent, LogicEvent_Beep, LogicEvent_CharacterEntered, LogicEvent_CharacterLeft, LogicEvent_Message, LogicEvent_RoomForceLeave, LogicEvent_RoomUpdate } from "../../../src/logicEvent";
-import { LogicBase } from "../../../src/logicBase";
 import { ConfigFile } from "../../config";
 
 export interface IAdminLogicSettings {
