+++ bin/hub/logic/loggingLogic.ts
@@ -2,11 +2,7 @@
  * Copied from the original bot hub, with permission.
  */
 
-import { logger } from "../../../src/api";
-import { API_Character } from "../../../src/apiCharacter";
-import { API_Connector } from "../../../src/apiConnector";
-import { LogicBase } from "../../../src/logicBase";
-import { AnyCharacterEvent, BC_Server_AccountBeep, BC_Server_ChatRoomMessage } from "../../../src/logicEvent";
+import { logger, API_Character, API_Connector, LogicBase, AnyCharacterEvent, BC_Server_AccountBeep, BC_Server_ChatRoomMessage } from "bc-bot";
 
 export class LoggingLogic extends LogicBase {
 	/**
