+++ bin/hub/utils.ts
@@ -47,4 +47,3 @@ function anyToString(data: unknown): string {
 		}) ?? "undefined"
 	);
 }
-
