+++ package.json
@@ -1,20 +1,21 @@
 {
     "name": "ropeybot",
     "license": "Apache-2.0",
+    "private": "true",
     "scripts": {
         "types": "tsc",
-        "bundle": "esbuild bin/main.ts --bundle --platform=node --outfile=out/bundle.js --sourcemap",
+        "bundle": "esbuild bin/main.ts --bundle --platform=node --outfile=dist/bundle.js --sourcemap",
         "start": "tsx bin/main.ts",
         "docker": "pnpm run bundle && docker build -t ropeybot --no-cache ."
     },
     "dependencies": {
+        "bc-bot": "link:src",
         "bc-stubs": "^116.0.0",
         "lodash": "^4.17.21",
         "lz-string": "^1.5.0",
         "mongodb": "^6.16.0",
         "prettier": "^3.5.3",
-        "prom-client": "^14.2.0",
-        "socket.io-client": "^4.8.1"
+        "prom-client": "^14.2.0"
     },
     "devDependencies": {
         "@tsconfig/node18": "^18.2.4",
