+++ src/package.json
@@ -1,0 +1,25 @@
+{
+    "name": "bc-bot",
+    "version": "0.0.1",
+    "license": "Apache-2.0",
+    "description": "",
+    "main": "dist/index.js",
+    "type": "module",
+    "types": "dist/index.d.ts",
+    "scripts": {
+        "compile": "rm -rf dist/* && tsc -p tsconfig.json"
+    },
+    "dependencies": {
+        "bc-stubs": "^116.0.0",
+        "lz-string": "^1.5.0",
+        "prettier": "^3.5.3",
+        "socket.io-client": "^4.8.1"
+    },
+    "devDependencies": {
+        "@tsconfig/node18": "^18.2.4",
+        "@types/node": "^20.17.50",
+        "esbuild": "^0.19.12",
+        "tsx": "^4.19.4",
+        "typescript": "^5.8.3"
+    }
+}
