+++ src/tsconfig.json
@@ -1,18 +1,20 @@
 {
   "compilerOptions": {
       "allowJs": true,
-      "noEmit": true,
+      "noEmit": false,
       "strict": false,
       "module": "commonjs",
       "moduleResolution": "node",
-      "esModuleInterop": true
+      "esModuleInterop": true,
+      "declaration": true,
+      "outDir": "./dist/",
   },
   "extends": "@tsconfig/node18/tsconfig.json",
   "include": [
       "node_modules/bc-stubs/bc/NativeDeclarations/Typedef.d.ts",
       "node_modules/bc-stubs/bc/NativeDeclarations/Messages.d.ts",
       "node_modules/bc-stubs/bc/NativeDeclarations/Female3DCG_Types.d.ts",
-      "bin/**/*",
+      "**/*",
   ],
   "ts-node": {
     "files": true
